Delegate Bukkit command registering logic to `CommandRegistrationStrategy`
Just a rough example implementation

`NMS#getResourceDispatcher` replaced by `NMS#createCommandRegistrationStrategy`, which creates a `SpigotCommandRegistration` implementation using the resources dispatcher

Alternate fix for https://github.com/JorelAli/CommandAPI/issues/554. If we're on Paper-1.20.6-65, a `PaperCommandRegistration` implementation is used instead. This can be used to properly handle the internal changes made by https://github.com/PaperMC/Paper/pull/8235.
