+++ commandapi-platforms/commandapi-bukkit/commandapi-bukkit-nms/commandapi-bukkit-nms-common/src/main/java/dev/jorel/commandapi/nms/NMS_Common.java
@@ -28,7 +28,7 @@ import com.mojang.brigadier.suggestion.SuggestionProvider;
 import com.mojang.brigadier.tree.CommandNode;
 import dev.jorel.commandapi.CommandAPIBukkit;
 import dev.jorel.commandapi.CommandAPIHandler;
-import dev.jorel.commandapi.SafeVarHandle;
+import dev.jorel.commandapi.CommandRegistrationStrategy;
 import dev.jorel.commandapi.arguments.ArgumentSubType;
 import dev.jorel.commandapi.arguments.SuggestionProviders;
 import dev.jorel.commandapi.commandsenders.AbstractCommandSender;
@@ -49,7 +49,6 @@ import net.minecraft.commands.arguments.coordinates.*;
 import net.minecraft.commands.arguments.item.FunctionArgument;
 import net.minecraft.network.chat.Component.Serializer;
 import net.minecraft.resources.ResourceLocation;
-import net.minecraft.server.MinecraftServer;
 import net.minecraft.world.phys.Vec2;
 import org.bukkit.*;
 import org.bukkit.block.Block;
@@ -70,7 +69,6 @@ import org.bukkit.scoreboard.Team;
 
 import java.io.File;
 import java.io.IOException;
-import java.lang.reflect.Field;
 import java.util.*;
 import java.util.function.Function;
 import java.util.function.Predicate;
@@ -93,19 +91,6 @@ import static dev.jorel.commandapi.preprocessor.Unimplemented.REASON.*;
  * (introducing another NMS_Common module perhaps?
  */
 public abstract class NMS_Common extends CommandAPIBukkit<CommandSourceStack> {
-
-	private static final Field commandDispatcher;
-
-	static {
-		Field temporary;
-		try {
-			temporary = MinecraftServer.class.getDeclaredField("vanillaCommandDispatcher");
-		} catch (Exception e) {
-			temporary = null;
-		}
-		commandDispatcher = temporary;
-	}
-
 	private static NamespacedKey fromResourceLocation(ResourceLocation key) {
 		return NamespacedKey.fromString(key.getNamespace() + ":" + key.getPath());
 	}
@@ -363,18 +348,8 @@ public abstract class NMS_Common extends CommandAPIBukkit<CommandSourceStack> {
 	public abstract BlockData getBlockState(CommandContext<CommandSourceStack> cmdCtx, String key);
 
 	@Override
-	public final CommandDispatcher<CommandSourceStack> getBrigadierDispatcher() {
-		if (commandDispatcher != null) {
-			return this.<MinecraftServer>getMinecraftServer().vanillaCommandDispatcher.getDispatcher();
-		} else {
-			// Hoping this actually acts as a hotfix and commands at least register with Paper 1.20.6 build 65
-			return getResourcesDispatcher();
-		}
-	}
-
-	@Override
 	@Unimplemented(because = NAME_CHANGED, info = "MinecraftServer#getCommands() obfuscated differently across multiple versions")
-	public abstract CommandDispatcher<CommandSourceStack> getResourcesDispatcher();
+	public abstract CommandRegistrationStrategy<CommandSourceStack> createCommandRegistrationStrategy();
 
 	@Override
 	@Overridden(in = "1.20.5", because = "Serializer.toJson now needs a Provider")
